+++ README.md
@@ -6,4 +6,4 @@ Tarpit is a web API for Deno modeled after ASP.NET / Spring Boot
 
 ### Why does this exist?
 
-While other MVC/web API libraries and frameworks provide a variety of features, many of which I find to be useful/desirable on their own, I've often found that there's one or two almost deal-breaking features that make me wish I was using a different one. In short, Tarpit is basically my wishlist come to life for a web API.
+// TODO
