+++ lib/decorators/request-body.ts
@@ -1,6 +1,6 @@
-import { _BODY_DECORATOR_META_KEY, BodyMetadata } from '../metadata.ts';
+import { _BODY_DECORATOR_META_KEY, BodyMetadata } from '../main/metadata.ts';
 import { Validator, ObjValidator, StringValidator, ArrayValidator } from '../validation/mod.ts';
-import { Logger } from '../logger.ts';
+import { Logger } from '../main/logger.ts';
 import 'https://deno.land/x/reflection@0.0.2/mod.ts';
 
 /**
