+++ lib/decorators/request-mapping.ts
@@ -1,4 +1,4 @@
-import { _ENDPOINT_DECORATOR_META_KEY, DecoratorRouteMetadata } from '../metadata.ts';
+import { _ENDPOINT_DECORATOR_META_KEY, DecoratorRouteMetadata } from '../main/metadata.ts';
 import 'https://deno.land/x/reflection@0.0.2/mod.ts';
 
 /**
