+++ lib/main/config.ts
