+++ lib/main/logger.ts
