+++ lib/main/mod.ts
@@ -1,0 +1,4 @@
+export * from './controller.ts';
+export * from './tarpit.ts';
+export * from './logger.ts';
+export type { ServerConfig } from './config.ts';
