+++ lib/main/tarpit.ts
