+++ mod.ts
@@ -1,8 +1,5 @@
+export * from './lib/main/mod.ts';
 export * from './lib/decorators/mod.ts';
-export * from './lib/controller.ts';
-export * from './lib/tarpit.ts';
 export * from './lib/request/mod.ts';
 export * from './lib/response/mod.ts';
 export * from './lib/validation/mod.ts';
-
-export type { ServerConfig } from './lib/config.ts';
